refactor(components): migrate Modal and ModalProvider to TypeScript
