+++ package.json
@@ -112,8 +112,11 @@
     "@types/jest-axe": "^3.2.2",
     "@types/jscodeshift": "^0.7.1",
     "@types/lodash": "^4.14.149",
+    "@types/no-scroll": "^2.1.0",
     "@types/react": "^16.9.32",
     "@types/react-dom": "^16.9.6",
+    "@types/react-modal": "^3.10.5",
+    "@types/recompose": "^0.30.7",
     "audit-ci": "^2.1.0",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^25.0.0",
