+++ src/components/Modal/Modal.docs.mdx
@@ -1,9 +1,5 @@
 import { Status, Props, Story } from '../../../.storybook/components';
-import {
-  ModalWrapper,
-  ModalHeader,
-  ModalFooter
-} from '../../../src/components/Modal';
+import { ModalWrapper, ModalHeader, ModalFooter } from '.';
 
 # Modal
 
@@ -14,7 +10,6 @@ modal should be related to completing one single task. Modals are heavy UI
 elements which obscure the primary user interface — avoid them where possible.
 
 <Story id="components-modal--base" />
-
 <Props of={ModalWrapper} />
 <Props of={ModalHeader} />
 <Props of={ModalFooter} />
